--- src/@types/clendar.d.ts
@@ -1,8 +1,0 @@
-import type { Dayjs } from "dayjs";
-
-import type { TSchedule } from "./schedule";
-
-export type CalendarProps = {
-  date: Dayjs;
-  schedules?: TSchedule[];
-};
