--- src/components/calendar/CalendarRangePicker.module.scss
@@ -1,42 +1,0 @@
-.wrapper {
-  width: 100%;
-  height: 100%;
-  display: flex;
-  flex-direction: column;
-  gap: 8px;
-  .toolbar {
-    display: flex;
-    justify-content: flex-end;
-    align-items: center;
-    .toggle {
-      background-color: #f5f5f5;
-      border: solid 1px #e0e0e0;
-      border-radius: 24px;
-      &.active {
-        .on {
-          background-color: #006a71;
-          .icon {
-            fill: white;
-          }
-        }
-        .off {
-          background-color: transparent;
-          .icon {
-            fill: child;
-          }
-        }
-      }
-      .on {
-        background-color: transparent;
-      }
-      &:not(.active) {
-        .off {
-          background-color: #c50000;
-          .icon {
-            fill: white;
-          }
-        }
-      }
-    }
-  }
-}
