--- src/components/calendar/index.tsx
@@ -1,1 +1,0 @@
-export * from "./CalendarRangePicker";
