--- src/components/calendar/static/week.ts
@@ -1,9 +1,0 @@
-export const DayOfWeekName = [
-  "日",
-  "月",
-  "火",
-  "水",
-  "木",
-  "金",
-  "土",
-] as const;
