--- src/components/calendar/timeline-wrapper/const.ts
@@ -1,2 +1,0 @@
-const timelineRange: void[] = [...(Array(24) as void[])];
-export { timelineRange };
