--- src/components/calendar/timeline-wrapper/current-time-indicator.module.scss
@@ -1,19 +1,0 @@
-.wrapper {
-  position: absolute;
-  background: red;
-  height: 2px;
-  left: 0;
-  width: 100%;
-  &:before {
-    content: "";
-    display: block;
-    width: 12px;
-    height: 12px;
-    border-radius: 6px;
-    position: absolute;
-    background: red;
-    left: 0;
-    top: 50%;
-    transform: translate(-50%, -50%);
-  }
-}
