--- src/components/calendar/timeline-wrapper/index.tsx
@@ -1,1 +1,0 @@
-export * from "./wrapper";
