--- src/components/calendar/timeline-wrapper/indicator.module.scss
@@ -1,40 +1,0 @@
-@import "../variable";
-
-.wrapper {
-  display: flex;
-  flex-direction: column;
-  height: 100%;
-  width: 32px;
-  flex-shrink: 0;
-  .container {
-    flex: 1;
-    display: flex;
-    flex-direction: column;
-    .item {
-      flex: 1;
-      position: relative;
-      .text {
-        position: absolute;
-        top: 0;
-        transform: translateY(-50%);
-        left: 0;
-        width: 100%;
-        text-align: center;
-      }
-      .border {
-        position: absolute;
-        top: 0;
-        right: 0;
-        width: 4px;
-        border-top: solid 1px $border-color;
-      }
-      &:first-child {
-        border: none;
-        .text,
-        .border {
-          display: none;
-        }
-      }
-    }
-  }
-}
