--- src/components/calendar/timeline-wrapper/scroll-detector.module.scss
@@ -1,40 +1,0 @@
-.wrapper {
-  width: 10%;
-  height: 100%;
-  position: absolute;
-  top: 0;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  &.left {
-    left: 0;
-    &:hover {
-      background: linear-gradient(
-        90deg,
-        rgba(0, 0, 0, 0.1) 0%,
-        rgba(0, 0, 0, 0) 100%
-      );
-    }
-  }
-  &.right {
-    right: 0;
-    &:hover {
-      background: linear-gradient(
-        270deg,
-        rgba(0, 0, 0, 0.1) 0%,
-        rgba(0, 0, 0, 0) 100%
-      );
-    }
-    .icon {
-      transform: rotate(180deg);
-    }
-  }
-  .icon {
-    display: none;
-  }
-  &:hover {
-    .icon {
-      display: block;
-    }
-  }
-}
