--- src/components/calendar/timeline-wrapper/selecting-indicator.module.scss
@@ -1,7 +1,0 @@
-.background {
-  position: absolute;
-  top: 0;
-  height: 100px;
-  pointer-events: none;
-  z-index: 1;
-}
