--- src/components/calendar/variable.scss
@@ -1,6 +1,0 @@
-$date-border: 2px;
-
-$border-color: #888;
-
-$badge-color: #55f;
-$badge-text-color: #fff;
