--- src/libraries/isTouchDevice.ts
@@ -1,5 +1,0 @@
-const isTouchDevice = (): boolean => {
-  return "ontouchstart" in window || navigator.maxTouchPoints > 0;
-};
-
-export { isTouchDevice };
