--- src/styles/create.module.scss
@@ -1,8 +1,0 @@
-.wrapper {
-  .rangePicker {
-    display: block;
-    &.invisible {
-      display: none;
-    }
-  }
-}
